+++ tests/provider/dwd/radar/test_api_invalid.py
@@ -5,7 +5,7 @@ from datetime import datetime
 
 import pytest
 
-from wetterdienst.exceptions import InvalidEnumeration
+from wetterdienst.exceptions import InvalidEnumerationError
 from wetterdienst.provider.dwd.radar import DwdRadarParameter, DwdRadarValues
 from wetterdienst.provider.dwd.radar.metadata import (
     DwdRadarDataFormat,
@@ -21,8 +21,7 @@ def test_radar_request_site_historic_pe_wrong_parameters(default_settings):
     Verify acquisition of radar/site/PE_ECHO_TOP data croaks
     when omitting RadarDataFormat.
     """
-
-    with pytest.raises(ValueError) as excinfo:
+    with pytest.raises(ValueError) as exec_info:
         request = DwdRadarValues(
             parameter=DwdRadarParameter.PE_ECHO_TOP,
             site=DwdRadarSite.BOO,
@@ -30,9 +29,7 @@ def test_radar_request_site_historic_pe_wrong_parameters(default_settings):
             settings=default_settings,
         )
         next(request.query())
-
-    assert excinfo.typename == "ValueError"
-    assert str(excinfo.value).startswith("Argument 'format' is missing")
+    assert exec_info.match("Argument 'format' is missing")
 
 
 def test_radar_request_site_historic_pe_future(
@@ -64,8 +61,7 @@ def test_radar_request_site_latest_sweep_pcp_v_hdf5(default_settings):
     """
     Verify requesting latest HDF5 data croaks.
     """
-
-    with pytest.raises(ValueError) as excinfo:
+    with pytest.raises(ValueError) as exec_info:
         request = DwdRadarValues(
             parameter=DwdRadarParameter.SWEEP_PCP_VELOCITY_H,
             site=DwdRadarSite.BOO,
@@ -75,52 +71,41 @@ def test_radar_request_site_latest_sweep_pcp_v_hdf5(default_settings):
         )
 
         list(request.query())
-
-    assert excinfo.typename == "ValueError"
-    assert str(excinfo.value).startswith("HDF5 data has no '-latest-' files")
+    assert exec_info.match("HDF5 data has no '-latest-' files")
 
 
 def test_radar_request_site_latest_sweep_pcp_v_hdf5_wrong_parameters(default_settings):
     """
     Verify requesting HDF5 data without RadarDataFormat croaks.
     """
-
-    with pytest.raises(ValueError) as excinfo:
+    with pytest.raises(ValueError) as exec_info:
         request = DwdRadarValues(
             parameter=DwdRadarParameter.SWEEP_PCP_VELOCITY_H,
             site=DwdRadarSite.BOO,
             start_date=DwdRadarDate.CURRENT,
             settings=default_settings,
         )
-
         list(request.query())
-
-    assert excinfo.typename == "ValueError"
-    assert str(excinfo.value).startswith("Argument 'format' is missing")
+    assert exec_info.match("Argument 'format' is missing")
 
 
 def test_radar_request_site_without_site(default_settings):
     """
     Verify requesting site data without site croaks.
     """
-
-    with pytest.raises(ValueError) as excinfo:
+    with pytest.raises(ValueError) as exec_info:
         request = DwdRadarValues(
             parameter=DwdRadarParameter.SWEEP_PCP_VELOCITY_H, start_date=DwdRadarDate.LATEST, settings=default_settings
         )
-
         list(request.query())
-
-    assert excinfo.typename == "ValueError"
-    assert str(excinfo.value).startswith("Argument 'site' is missing")
+    assert exec_info.match("Argument 'site' is missing")
 
 
 def test_radar_request_hdf5_without_subset(default_settings):
     """
     Verify requesting HDF5 data without "subset" croaks.
     """
-
-    with pytest.raises(ValueError) as excinfo:
+    with pytest.raises(ValueError) as exec_info:
         request = DwdRadarValues(
             parameter=DwdRadarParameter.SWEEP_PCP_VELOCITY_H,
             site=DwdRadarSite.BOO,
@@ -128,11 +113,8 @@ def test_radar_request_hdf5_without_subset(default_settings):
             start_date=DwdRadarDate.MOST_RECENT,
             settings=default_settings,
         )
-
         list(request.query())
-
-    assert excinfo.typename == "ValueError"
-    assert str(excinfo.value).startswith("Argument 'subset' is missing")
+    assert exec_info.match("Argument 'subset' is missing")
 
 
 @pytest.mark.remote
@@ -147,19 +129,15 @@ def test_radar_request_radolan_cdc_latest(time_resolution, default_settings):
     """
     Verify requesting latest RADOLAN_CDC croaks.
     """
-
-    with pytest.raises(ValueError) as excinfo:
+    with pytest.raises(ValueError) as exec_info:
         request = DwdRadarValues(
             parameter=DwdRadarParameter.RADOLAN_CDC,
             resolution=time_resolution,
             start_date=DwdRadarDate.LATEST,
             settings=default_settings,
         )
-
         list(request.query())
-
-    assert excinfo.typename == "ValueError"
-    assert str(excinfo.value).startswith("RADOLAN_CDC data has no '-latest-' files")
+    assert exec_info.match("RADOLAN_CDC data has no '-latest-' files")
 
 
 def test_radar_request_radolan_cdc_invalid_time_resolution(default_settings):
@@ -167,7 +145,7 @@ def test_radar_request_radolan_cdc_invalid_time_resolution(default_settings):
     Verify requesting 1-minute RADOLAN_CDC croaks.
     """
 
-    with pytest.raises(InvalidEnumeration):
+    with pytest.raises(InvalidEnumerationError):
         DwdRadarValues(
             parameter=DwdRadarParameter.RADOLAN_CDC,
             resolution="minute_1",