+++ wetterdienst/core/process.py
@@ -7,7 +7,7 @@ from typing import Optional, Tuple
 import polars as pl
 from backports.datetime_fromisoformat import MonkeyPatch
 
-from wetterdienst.exceptions import InvalidTimeInterval
+from wetterdienst.exceptions import InvalidTimeIntervalError
 from wetterdienst.metadata.columns import Columns
 from wetterdienst.metadata.resolution import Resolution
 from wetterdienst.util.datetime import mktimerange, parse_date
@@ -18,7 +18,7 @@ MonkeyPatch.patch_fromisoformat()
 def create_date_range(date: str, resolution: Resolution) -> Tuple[Optional[dt.datetime], Optional[dt.datetime]]:
     if "/" in date:
         if date.count("/") >= 2:
-            raise InvalidTimeInterval("Invalid ISO 8601 time interval")
+            raise InvalidTimeIntervalError("Invalid ISO 8601 time interval")
 
         date_from, date_to = date.split("/")
         date_from = parse_date(date_from)
@@ -68,7 +68,7 @@ def filter_by_date(df: pl.DataFrame, date: str) -> pl.DataFrame:
     # Filter by date interval.
     if "/" in date:
         if date.count("/") >= 2:
-            raise InvalidTimeInterval("Invalid ISO 8601 time interval")
+            raise InvalidTimeIntervalError("Invalid ISO 8601 time interval")
 
         date_from, date_to = date.split("/")
         date_from = parse_date(date_from)